made the base header and trying to make home page
