+++ SP23-P03-Mobile/App.js
@@ -1,15 +1,20 @@
 import { StatusBar } from 'expo-status-bar';
 import { StyleSheet, Text, View } from 'react-native';
-import { HeaderApp } from './App_src/Header';
+import { HeaderApp } from './App_src/Compenents/Header';
+import {NavigationContainer} from '@react-navigation/native';
+import {createNativeStackNavigator} from '@react-navigation/native-stack';
+import {HomeScreen} from './App_src/Pages/Homepage' ;
+
+
+const Stack = createNativeStackNavigator();
 
 export default function App() {
   return (
 
     <View style={styles.container}>
       <HeaderApp/>
-      <View style ={styles.content}>
-      <Text>Open up App.js to start working on your app!</Text>
-      <Text>This is kyle tran!</Text>
+        <View style ={styles.content}>
+          <Text>Words</Text>
       <StatusBar style="auto" />
       </View>
     </View>
