+++ SP23-P03-Mobile/App_src/Pages/Homepage.js
@@ -1,0 +1,17 @@
+import React from "react";
+import { View } from "react-native";
+
+
+const HomeScreen = ({navigation}) => {
+    return (
+        <View>
+            <Button
+            title="Go to Jane's profile"
+            onPress={() =>
+            null
+            }
+            />
+      </View>
+    );
+  };
+
